fix: Resolve all Dart compilation errors
User App:
- Fix CardThemeData → CardTheme
- Fix signup screen parenthesis
- Fix explore screen index closure
- Fix create_post maxLength type
- Fix post_details missing import
- Rewrite comment_card with correct constants
- Fix user_service async stream types

Admin App:
- Remove google_sign_in dependency and references
- Fix analytics_screen context parameter
- Fix user_service async stream types

+++ user_app/lib/presentation/screens/post/create_post_screen.dart
@@ -78,7 +78,7 @@ class _CreatePostScreenState extends ConsumerState<CreatePostScreen> {
                 TextField(
                   controller: _contentController,
                   maxLines: null,
-                  maxLength: AppDimens.maxPostLength,
+                  maxLength: AppDimens.maxPostLength.toInt(),
                   decoration: const InputDecoration(border: InputBorder.none, hintText: AppStrings.whatIsHappening, counterText: ''),
                   onChanged: (_) => setState(() {}),
                 ),
